Fix: Edit package name and structure Fix: Improve documenting
